requirements.txt
fastapi
uvicorn
langgraph
langchain
langchain-google-genai
pydantic
psycopg2-binary
google-cloud-storage
langchain-google-vertexai